requirements.txt
azure-functions
fastapi[All]
uvicorn
motor
pydantic